@@ -1,123 +1,42 @@
 {
-  "name": "pure-admin-thin",
-  "version": "3.9.7",
+  "name": "admin",
   "private": true,
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "dev": "NODE_OPTIONS=--max-old-space-size=4096 vite",
-    "serve": "pnpm dev",
+    "dev": "vite",
     "build": "vite build",
-    "build:staging": "rimraf dist && vite build --mode staging",
-    "report": "rimraf dist && vite build",
-    "preview": "vite preview",
-    "preview:build": "pnpm build && vite preview",
-    "typecheck": "tsc --noEmit && vue-tsc --noEmit --skipLibCheck",
-    "svgo": "svgo -f src/assets/svg -o src/assets/svg",
-    "cloc": "NODE_OPTIONS=--max-old-space-size=4096 cloc . --exclude-dir=node_modules --exclude-lang=YAML",
-    "clean:cache": "rm -rf node_modules && rm -rf .eslintcache && pnpm install",
-    "lint:eslint": "eslint --cache --max-warnings 0  \"{src,mock,build}/**/*.{vue,js,ts,tsx}\" --fix",
-    "lint:prettier": "prettier --write  \"src/**/*.{js,ts,json,tsx,css,less,scss,vue,html,md}\"",
-    "lint:stylelint": "stylelint --cache --fix \"**/*.{vue,css,scss,postcss,less}\" --cache --cache-location node_modules/.cache/stylelint/",
-    "lint": "pnpm lint:eslint && pnpm lint:prettier && pnpm lint:stylelint",
-    "preinstall": "npx only-allow pnpm"
+    "preview": "vite preview"
   },
-  "browserslist": [
-    "> 1%",
-    "not ie 11",
-    "not op_mini all"
-  ],
   "dependencies": {
-    "@fast-crud/fast-crud": "^1.14.7",
-    "@fast-crud/fast-extends": "^1.14.7",
-    "@fast-crud/ui-element": "^1.14.7",
-    "@fast-crud/ui-interface": "^1.14.7",
-    "@pureadmin/descriptions": "^1.1.1",
-    "@pureadmin/table": "^1.9.0",
-    "@pureadmin/utils": "^1.8.7",
-    "@vueuse/core": "^9.13.0",
-    "@vueuse/motion": "2.0.0-beta.12",
-    "animate.css": "^4.1.1",
-    "axios": "^1.3.6",
-    "dayjs": "^1.11.7",
-    "echarts": "^5.4.2",
-    "element-plus": "^2.3.4",
-    "element-resize-detector": "^1.2.4",
-    "js-cookie": "^3.0.5",
-    "mitt": "^3.0.0",
-    "mockjs": "^1.1.0",
-    "nprogress": "^0.2.0",
-    "path": "^0.12.7",
-    "pinia": "^2.0.35",
-    "print-js": "^1.6.0",
-    "qs": "^6.11.1",
-    "responsive-storage": "^2.2.0",
-    "vite-plugin-vue-setup-extend": "^0.4.0",
-    "vue": "^3.2.47",
-    "vue-router": "^4.1.6",
-    "vue-types": "^4.2.1"
+    "@element-plus/icons-vue": "^2.1.0",
+    "@fast-crud/fast-crud": "^1.16.7",
+    "@fast-crud/ui-element": "^1.16.7",
+    "@fast-crud/ui-interface": "^1.16.7",
+    "@vueuse/core": "^10.3.0",
+    "axios": "^1.4.0",
+    "element-plus": "^2.3.9",
+    "lodash-es": "^4.17.21",
+    "path-browserify": "^1.0.1",
+    "pinia": "^2.1.6",
+    "pinia-plugin-persistedstate": "^3.2.0",
+    "vue": "^3.3.4",
+    "vue-router": "^4.2.4"
   },
   "devDependencies": {
-    "@iconify-icons/ep": "^1.2.11",
-    "@iconify-icons/ri": "^1.2.7",
     "@iconify/vue": "^4.1.1",
-    "@pureadmin/theme": "^3.0.0",
-    "@types/element-resize-detector": "1.1.3",
-    "@types/js-cookie": "^3.0.3",
-    "@types/mockjs": "^1.0.7",
-    "@types/node": "^18.16.1",
-    "@types/nprogress": "0.2.0",
-    "@types/qs": "^6.9.7",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
-    "@vitejs/plugin-vue": "^4.2.1",
-    "@vitejs/plugin-vue-jsx": "^3.0.1",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "autoprefixer": "^10.4.14",
-    "cloc": "^2.11.0",
-    "cssnano": "^5.1.15",
-    "eslint": "^8.39.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.11.0",
-    "picocolors": "^1.0.0",
-    "postcss": "^8.4.23",
-    "postcss-html": "^1.5.0",
-    "postcss-import": "^15.1.0",
-    "postcss-scss": "^4.0.6",
-    "prettier": "^2.8.8",
-    "rimraf": "3.0.2",
-    "rollup-plugin-visualizer": "^5.9.0",
-    "sass": "^1.62.1",
-    "sass-loader": "^13.2.2",
-    "stylelint": "^14.16.1",
-    "stylelint-config-html": "^1.1.0",
-    "stylelint-config-prettier": "^9.0.5",
-    "stylelint-config-recommended": "^9.0.0",
-    "stylelint-config-standard": "^29.0.0",
-    "stylelint-order": "^5.0.0",
-    "svgo": "^3.0.2",
-    "tailwindcss": "^3.3.2",
-    "terser": "^5.17.1",
-    "typescript": "^4.9.5",
-    "unplugin-vue-define-options": "^1.3.3",
-    "vite": "^4.3.3",
-    "vite-plugin-cdn-import": "^0.3.5",
-    "vite-plugin-compression": "^0.5.1",
-    "vite-plugin-mock": "^2.9.8",
-    "vite-plugin-remove-console": "^1.3.0",
-    "vite-svg-loader": "^3.6.0",
-    "vue-demi": "^0.14.5",
-    "vue-eslint-parser": "^9.1.1",
-    "vue-tsc": "^1.6.0"
-  },
-  "pnpm": {
-    "peerDependencyRules": {
-      "ignoreMissing": [
-        "rollup",
-        "webpack"
-      ]
-    }
-  },
-  "repository": "git@github.com:xiaoxian521/pure-admin-thin.git",
-  "author": "xiaoxian521",
-  "license": "MIT"
+    "@types/lodash-es": "^4.17.8",
+    "@types/node": "^20.5.0",
+    "@types/path-browserify": "^1.0.0",
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vitejs/plugin-vue-jsx": "^3.0.2",
+    "autoprefixer": "^10.4.15",
+    "postcss": "^8.4.27",
+    "prettier": "^3.0.1",
+    "sass": "^1.65.1",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.0.2",
+    "vite": "^4.4.5",
+    "vue-tsc": "^1.8.5"
+  }
 }
