@@ -1,20 +1,8 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
-  darkMode: "class",
-  corePlugins: {
-    preflight: false
-  },
-  content: ["./index.html", "./src/**/*.{vue,js,ts,jsx,tsx}"],
+export default {
+  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx,vue}"],
   theme: {
-    extend: {
-      colors: {
-        bg_color: "var(--el-bg-color)",
-        primary: "var(--el-color-primary)",
-        primary_light_9: "var(--el-color-primary-light-9)",
-        text_color_primary: "var(--el-text-color-primary)",
-        text_color_regular: "var(--el-text-color-regular)",
-        text_color_disabled: "var(--el-text-color-disabled)"
-      }
-    }
-  }
+    extend: {},
+  },
+  plugins: [],
 };
