@@ -1,9 +1,23 @@
-import Crud from "@/api/crud";
-import { http } from "@/utils/http";
+import Crud, {QueryParams} from "@/api/crud";
+import { request } from "@/utils/request";
+import {
+  MenuSchemaFilters,
+  MenuSchemaTree,
+  PageMenuSchemaList,
+} from "@/api/menu/types";
 
 class Menu extends Crud {
-  async getMenuTree() {
-    return await http.get("/menu/tree");
+  async list(
+    query:QueryParams={},
+    filters: MenuSchemaFilters = {},
+  ): Promise<PageMenuSchemaList> {
+    return await super.list(query, filters);
+  }
+  async getMenuTree(): Promise<MenuSchemaTree> {
+    return await request({
+      method: "get",
+      url: "/menu/tree",
+    });
   }
 }
 
