@@ -7,15 +7,16 @@ export function useSpiderButtons(model: any) {
       text: "采集",
       type: "success",
       loading: spiderLoading,
+      // @ts-ignore
       click: ({ form }) => {
         spiderLoading.value = true;
         model
           .getSiteInfo(form.href)
-          .then(data => Object.assign(form, data))
+          .then((data: any) => Object.assign(form, data))
           .finally(() => {
             spiderLoading.value = false;
           });
-      }
-    }
+      },
+    },
   };
 }
