@@ -1,12 +1,14 @@
 // 引入fast-crud
-import { FastCrud, setLogger } from "@fast-crud/fast-crud";
+import { FastCrud, setLogger, useTypes } from "@fast-crud/fast-crud";
 import "@fast-crud/fast-crud/dist/style.css";
 
 // 请选择ui: element/ antdv /naive。三选一，不支持动态切换
 // element
 import ui from "@fast-crud/ui-element";
 import { App, Component, ref } from "vue";
-import { http } from "@/utils/http";
+// @ts-ignore
+import { request } from "@/utils/request";
+import { registerPlugins } from "./plugins";
 
 // 设置日志级别
 setLogger("error");
@@ -16,10 +18,11 @@ export function useFastCrud(app: App) {
   app.use(ui);
   // 然后安装FastCrud
   app.use(FastCrud, {
+    logger: { off: { tableColumns: false } },
     // 此处配置公共的dictRequest（字典请求）
     async dictRequest(config) {
       // @ts-ignore
-      return await http.request(config); //根据dict的url，异步返回一个字典数组
+      return await request(config); //根据dict的url，异步返回一个字典数组
     },
     //公共crud配置
     commonOptions() {
@@ -30,11 +33,21 @@ export function useFastCrud(app: App) {
           //请参考文档http://fast-crud.docmirror.cn/api/crud-options/request.html
           transformQuery: ({ page, form, sort }) => {
             //转换为你pageRequest所需要的请求参数结构
-            return {
-              page: page.currentPage,
-              pageSize: page.pageSize,
-              filters: form
+            const query = {
+              params: {
+                page: page.currentPage,
+                size: page.pageSize,
+                order_by: "",
+              },
+              filters: form,
             };
+            // debugger;
+            if (typeof sort?.asc === "boolean") {
+              // @ts-ignore
+              query.params.order_by = sort.asc ? sort.prop : `-${sort.prop}`;
+            }
+
+            return query;
           },
           transformRes: ({ res }) => {
             //将pageRequest的返回数据，转换为fast-crud所需要的格式
@@ -43,36 +56,46 @@ export function useFastCrud(app: App) {
               records: res.items,
               currentPage: res.page,
               pageSize: res.size,
-              total: res.total
+              total: res.total,
             };
             // return { ...res };
-          }
+          },
         },
         //你可以在此处配置你的其他crudOptions公共配置
+        table: {
+          editable: {
+            enabled: true,
+            mode: "free", //模式，free=自由编辑，row=行编辑
+            activeTrigger: "onClick", //激活编辑的触发方式，dblclick=双击，click=单击
+          },
+        },
         rowHandle: {
           align: "center",
           buttons: {
             view: {
               icon: "View",
-              text: ""
+              text: "",
             },
             edit: {
               icon: "edit",
-              text: ""
+              text: "",
             },
             remove: {
               icon: "delete",
-              text: ""
-            }
-          }
+              text: "",
+            },
+          },
         },
         pagination: {
-          pageSize: 10
+          pageSize: 10,
+          background: true,
+          // "hide-on-single-page": true,
+          "page-sizes": ["5", "10", "20", "50", "100"],
         },
         form: {
           col: {
-            span: 24
-          }
+            span: 24,
+          },
         },
         columns: {
           $checked: {
@@ -82,11 +105,21 @@ export function useFastCrud(app: App) {
               type: "selection",
               align: "center",
               width: "55px",
-              columnSetDisabled: true //禁止在列设置中选择
-            }
-          }
-        }
+              columnSetDisabled: true, //禁止在列设置中选择
+            },
+          },
+          id: {
+            title: "id",
+            form: { show: false }, // 表单配置
+            column: {
+              width: 70,
+              // show: false,
+            },
+          },
+        },
       };
-    }
+    },
   });
+  // 注册自定义组件
+  registerPlugins();
 }
