@@ -1,17 +1,18 @@
 import { defineStore } from "pinia";
 import menu from "@/api/menu";
+import { MenuSchemaList, MenuSchemaTree } from "@/api/menu/types";
 
 /**
  * 记录分类和所有数据
  */
 export const useMenuStore = defineStore("menu", {
   state: () => ({
-    menuList: [],
-    menuTree: []
+    menuList: new Array<MenuSchemaList>(),
+    menuTree: new Array<MenuSchemaTree>(),
   }),
   actions: {
     async getMenuList() {
-      const { items } = await menu.list(1, 100);
+      const { items } = await menu.list({ size: 100 });
       this.menuList = items;
       // return items?.map(item => item.id);
       return items;
@@ -19,6 +20,7 @@ export const useMenuStore = defineStore("menu", {
     async getMenuTree() {
       // @ts-ignore
       this.menuTree = await menu.getMenuTree();
-    }
-  }
+      return this.menuTree;
+    },
+  },
 });
