@@ -1,162 +1,166 @@
 <script setup lang="ts">
-import Motion from "./utils/motion";
-import { useRouter } from "vue-router";
-import { message } from "@/utils/message";
-import { loginRules } from "./utils/rule";
-import { initRouter } from "@/router/utils";
-import { useNav } from "@/layout/hooks/useNav";
+import { reactive, ref } from "vue";
+import { loginRules } from "./rules.ts";
 import type { FormInstance } from "element-plus";
-import { useLayout } from "@/layout/hooks/useLayout";
-import { useUserStoreHook } from "@/store/modules/user";
-import { bg, avatar, illustration } from "./utils/static";
-import { useRenderIcon } from "@/components/ReIcon/src/hooks";
-import { ref, reactive, toRaw, onMounted, onBeforeUnmount } from "vue";
-import { useDataThemeChange } from "@/layout/hooks/useDataThemeChange";
+import { useUserStore } from "@/store";
+import router from "@/router";
+import { ElMessage } from "element-plus";
 
-import dayIcon from "@/assets/svg/day.svg?component";
-import darkIcon from "@/assets/svg/dark.svg?component";
-import Lock from "@iconify-icons/ri/lock-fill";
-import User from "@iconify-icons/ri/user-3-fill";
-import { usePermissionStoreHook } from "@/store/modules/permission";
-
-defineOptions({
-  name: "Login"
-});
-const router = useRouter();
-const loading = ref(false);
-const ruleFormRef = ref<FormInstance>();
-
-const { initStorage } = useLayout();
-initStorage();
-
-const { dataTheme, dataThemeChange } = useDataThemeChange();
-dataThemeChange();
-const { title } = useNav();
-
-const ruleForm = reactive({
+const loginForm = reactive({
   username: "",
-  password: ""
+  password: "",
 });
+const userStore = useUserStore();
+const loginFormRef = ref<FormInstance>();
+const loading = ref(false);
 
-const onLogin = async (formEl: FormInstance | undefined) => {
-  loading.value = true;
-  if (!formEl) return;
-  await formEl.validate((valid, fields) => {
+function handleSubmit() {
+  loginFormRef.value?.validate((valid: boolean, fields) => {
     if (valid) {
-      useUserStoreHook()
-        .loginByUsername(ruleForm)
-        .then(res => {
-          usePermissionStoreHook().handleWholeMenus([]);
-          message("登录成功", { type: "success" });
-          router.push("/");
-          // 获取后端路由
-          // initRouter().then(() => {
-          //   router.push("/");
-          // });
+      loading.value = true;
+      userStore
+        .login(loginForm)
+        .then(() => {
+          ElMessage.success("登录成功");
+          router.push({
+            path: "/",
+          });
         })
         .finally(() => {
           loading.value = false;
         });
     } else {
-      loading.value = false;
-      return fields;
+      new Error("表单校验失败");
     }
   });
-};
-
-/** 使用公共函数，避免`removeEventListener`失效 */
-function onkeypress({ code }: KeyboardEvent) {
-  if (code === "Enter") {
-    onLogin(ruleFormRef.value);
-  }
 }
-
-onMounted(() => {
-  window.document.addEventListener("keypress", onkeypress);
-});
-
-onBeforeUnmount(() => {
-  window.document.removeEventListener("keypress", onkeypress);
-});
 </script>
 
 <template>
-  <div class="select-none">
-    <img :src="bg" class="wave" />
-    <div class="flex-c absolute right-5 top-3">
-      <!-- 主题 -->
-      <el-switch
-        v-model="dataTheme"
-        inline-prompt
-        :active-icon="dayIcon"
-        :inactive-icon="darkIcon"
-        @change="dataThemeChange"
-      />
-    </div>
-    <div class="login-container">
-      <div class="img">
-        <component :is="toRaw(illustration)" />
+  <div class="login-container">
+    <el-form
+      class="login-form"
+      ref="loginFormRef"
+      :model="loginForm"
+      :rules="loginRules"
+      @keyup.enter="handleSubmit"
+    >
+      <div class="title-container">
+        <h3 class="title">用户登录</h3>
       </div>
-      <div class="login-box">
-        <div class="login-form">
-          <avatar class="avatar" />
-          <Motion>
-            <h2 class="outline-none">{{ title }}</h2>
-          </Motion>
 
-          <el-form
-            ref="ruleFormRef"
-            :model="ruleForm"
-            :rules="loginRules"
-            size="large"
-          >
-            <Motion :delay="100">
-              <el-form-item prop="username">
-                <el-input
-                  clearable
-                  v-model="ruleForm.username"
-                  placeholder="账号"
-                  :prefix-icon="useRenderIcon(User)"
-                />
-              </el-form-item>
-            </Motion>
+      <el-form-item prop="username">
+        <el-input
+          v-model.trim="loginForm.username"
+          clearable
+          prefix-icon="user"
+          placeholder="请输入用户名"
+          type="text"
+        />
+      </el-form-item>
 
-            <Motion :delay="150">
-              <el-form-item prop="password">
-                <el-input
-                  clearable
-                  show-password
-                  v-model="ruleForm.password"
-                  placeholder="密码"
-                  :prefix-icon="useRenderIcon(Lock)"
-                />
-              </el-form-item>
-            </Motion>
-
-            <Motion :delay="250">
-              <el-button
-                class="w-full mt-4"
-                size="default"
-                type="primary"
-                :loading="loading"
-                @click="onLogin(ruleFormRef)"
-              >
-                登录
-              </el-button>
-            </Motion>
-          </el-form>
-        </div>
-      </div>
-    </div>
+      <el-form-item prop="password">
+        <el-input
+          v-model.trim="loginForm.password"
+          prefix-icon="Lock"
+          show-password
+          clearable
+          placeholder="请输入密码"
+          type="password"
+        />
+      </el-form-item>
+      <el-button type="primary" :loading="loading" @click.prevent="handleSubmit"
+        >登录
+      </el-button>
+    </el-form>
   </div>
 </template>
 
-<style scoped>
-@import url("@/style/login.css");
-</style>
-
 <style lang="scss" scoped>
-:deep(.el-input-group__append, .el-input-group__prepend) {
-  padding: 0;
+$bg: #2d3a4b;
+$dark_gray: #889aa4;
+$light_gray: #eee;
+$cursor: #fff;
+
+.login-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-height: 100%;
+  width: 100%;
+  background-color: $bg;
+  overflow: hidden;
+
+  .login-form {
+    position: relative;
+    width: 520px;
+    max-width: 80%;
+    margin: 0 auto;
+
+    .el-button {
+      width: 100%;
+      height: 50px;
+    }
+
+    .el-form-item {
+      padding: 0px 5px;
+      border: 1px solid rgba(255, 255, 255, 0.1);
+      background: rgba(0, 0, 0, 0.1);
+      border-radius: 5px;
+    }
+
+    :deep(.el-input) {
+      display: inline-block;
+      height: 47px;
+      //width: 100%;
+      .el-input__wrapper {
+        background-color: unset;
+        box-shadow: unset;
+        width: 100%;
+      }
+
+      input {
+        width: 100%;
+        border: 0px;
+        border-radius: 0px;
+        padding: 12px 5px 12px 15px;
+        color: $light_gray;
+        height: 47px;
+        caret-color: $cursor;
+      }
+    }
+  }
+
+  .tips {
+    font-size: 16px;
+    line-height: 28px;
+    color: #fff;
+    margin-bottom: 10px;
+
+    span {
+      &:first-of-type {
+        margin-right: 16px;
+      }
+    }
+  }
+
+  .svg-container {
+    padding: 6px 5px 6px 15px;
+    color: $dark_gray;
+    vertical-align: middle;
+    display: inline-block;
+  }
+
+  .title-container {
+    position: relative;
+
+    .title {
+      font-size: 26px;
+      color: $light_gray;
+      margin: 0px auto 40px auto;
+      text-align: center;
+      font-weight: bold;
+    }
+  }
 }
 </style>
