@@ -1,36 +1,99 @@
 <template>
-  <fs-page>
-    <fs-crud ref="crudRef" v-bind="crudBinding" />
-  </fs-page>
+  <m-table :crud-options="crudOptions" :model="menuModel"></m-table>
 </template>
 
 <script lang="ts" setup>
-import { ref, onMounted } from "vue";
-import { useCrud, useExpose } from "@fast-crud/fast-crud";
-import createCrudOptions from "./crud";
+import MTable from "@/components/table/index.vue";
+import { CrudOptions, dict } from "@fast-crud/fast-crud";
+import { useSpiderButtons } from "@/hooks/spider";
+import { UseIconForm } from "@/hooks/icon";
+import menuModel from "@/api/menu";
 
-// crud组件的ref
-const crudRef = ref();
-// crud 配置的ref
-const crudBinding = ref();
-// 暴露的方法
-const { crudExpose } = useExpose({ crudRef, crudBinding });
-// 你的crud配置
-const { crudOptions } = createCrudOptions(crudExpose);
-// eslint-disable-next-line @typescript-eslint/no-unused-vars,no-unused-vars
-// 初始化crud配置
-// 此处传入权限前缀进行通用按钮权限设置，会通过commonOptions去设置actionbar和rowHandle的按钮的show属性
-// 更多关于按钮权限的源代码设置，请参考 ./src/plugin/fast-crud/index.js （75-77行）
-const { resetCrudOptions } = useCrud({
-  crudExpose,
-  crudOptions,
-  context: {}
-});
-// 你可以调用此方法，重新初始化crud配置
-// resetCrudOptions(options)
-// 页面打开后获取列表数据
-onMounted(async () => {
-  await crudExpose.doRefresh();
-  // const resp = await permission.getPermissionTree();
-});
+const iconForm = UseIconForm(true);
+const { spider } = useSpiderButtons(menuModel);
+
+const crudOptions: CrudOptions = {
+  form: {
+    wrapper: {
+      buttons: {
+        spider,
+      },
+    },
+  },
+  request: {
+    pageRequest: async () => {
+      const data = await menuModel.getMenuTree();
+      return {
+        items: data,
+        page: 1,
+        pages: 1,
+        size: 10,
+        total: 10,
+      };
+    },
+  },
+  columns: {
+    id: {
+      column: {
+        width: 100,
+      },
+    },
+    title: {
+      title: "标题",
+      type: "text",
+      search: { show: true }, // 开启查询
+      form: {
+        rules: [{ required: true, message: "请输入标题" }],
+      },
+      column: {
+        width: 180,
+      },
+    },
+    ...iconForm,
+    parent_id: {
+      title: "父级菜单",
+      type: "dict-select",
+      column: {
+        show: false,
+      },
+      form: {
+        col: { span: 12 },
+      },
+      dict: dict({
+        getData: async () => {
+          const { items } = await menuModel.list(
+            { size: 100 },
+            { parent_id: 0 },
+          );
+          return items;
+        },
+        label: "title",
+        value: "id",
+        immediate: false,
+        // data: [{ label: "aa", value: "bb" }]
+      }),
+    },
+    order: {
+      title: "排序",
+      type: "number",
+      column: {
+        showTitle: "值越小越靠前",
+      },
+      form: {
+        col: { span: 12 },
+      },
+    },
+    create_time: {
+      title: "创建时间",
+      type: "datetime",
+      form: { show: false, submit: false }, // 表单配置
+      viewForm: {
+        show: true,
+      },
+      column: {
+        width: 180,
+      },
+    },
+  },
+};
 </script>
