@@ -1,12 +1,12 @@
 from fastapi import Depends
 
-from core.crud import ModelCrud, BaseApiOut
+from fastapi_tortoise_crud import ModelCrud, BaseApiOut
 from models import Friend
 from ..links.utils import get_site_info
-from core.auth import get_current_user
+from auth.auth import get_current_user
 
 friend_router = ModelCrud(Friend,
-                          schema_filters=Friend.schema_filters(include=('title',))
+                          schema_filters=Friend.schema_filters(include=('title', ))
                           ).register_crud(
     depends_create=[Depends(get_current_user)],
     depends_update=[Depends(get_current_user)]
