@@ -15,9 +15,9 @@ class MenuSchemaList(Menu.schema_list(exclude=('parent',))):
 
 
 class MenuSchemaUpdate(Menu.schema_update(exclude=('parent_id',))):
-    parent_id: int | str | None
+    parent_id: int | str | None = None
     pass
 
 
 class MenuSchemaFilters(Menu.schema_filters(include=('title',))):
-    parent_id: int | str | None
+    parent_id: int | str | None = None
