@@ -1,4 +1,4 @@
-from .user import User
+from .system import User
 from .menu import Menu
 from .links import Links
 from .site import Site
