@@ -1,4 +1,4 @@
-from core.db import BaseModel
+from fastapi_tortoise_crud import BaseModel
 from tortoise import fields
 
 
@@ -10,3 +10,4 @@ class Menu(BaseModel):
     order = fields.IntField(default=0, description='值越大越靠前')
     color = fields.CharField(7, null=True, description='图标颜色hex')
     links = fields.ManyToManyField("models.Links", related_name="menus", null=True)
+    is_vip = fields.BooleanField(default=False, description='是否是vip菜单，登录后才能查看')
