@@ -1,16 +1,22 @@
-import { request } from "@/utils/service";
+import { request } from "@/utils/request";
 
-interface QueryParams {
+export interface QueryParams {
   page?: number;
-  pageSize?: number;
+  size?: number;
   order_by?: string;
 }
 
 class Crud {
   protected readonly baseUrl: string;
+  protected defaultQueryParams: QueryParams;
 
   constructor(baseUrl: string = "") {
     this.baseUrl = baseUrl || this.constructor.name.toLowerCase();
+    this.defaultQueryParams = {
+      page: 1,
+      size: 10,
+      order_by: "-create_time",
+    };
   }
 
   /**
@@ -18,21 +24,26 @@ class Crud {
    * @param query
    * @param filters
    */
-  async list(
-    query: QueryParams = {
-      page: 1,
-      pageSize: 10,
-      order_by: "-create_time",
-    },
-    filters = {},
-  ): Promise<any> {
+  async list(query: QueryParams = {}, filters = {}): Promise<any> {
+    const queryParams = Object.assign(this.defaultQueryParams, query);
     return await request({
       method: "post",
       url: `${this.baseUrl}/list`,
       params: {
-        ...query,
+        ...queryParams,
       },
       data: filters,
+    });
+  }
+
+  /**
+   *  获取详情
+   * @param id
+   */
+  async read(id: number): Promise<any> {
+    return await request({
+      url: `${this.baseUrl}/read/${id}`,
+      method: "get",
     });
   }
 
@@ -83,6 +94,29 @@ class Crud {
       method: "delete",
     });
   }
+}
+
+export function useRequest(model: Crud) {
+  const pageRequest = async (query: any) => {
+    return await model.list(query.params, query.filters);
+  };
+  const editRequest = async ({ form, row }: any) => {
+    form.id = row.id;
+    return await model.update(row.id, form);
+  };
+  const delRequest = async ({ row }: any) => {
+    return await model.delete(row.id);
+  };
+
+  const addRequest = async ({ form }: any) => {
+    return await model.create(form);
+  };
+  return {
+    pageRequest,
+    addRequest,
+    editRequest,
+    delRequest,
+  };
 }
 
 export default Crud;
