@@ -25,26 +25,28 @@ function gotoList(item: any) {
 
 <template>
   <!--  如果有子类-->
-  <el-sub-menu
-    :index="item.title"
-    v-if="item?.children && item.children.length > 0"
-  >
-    <template #title>
-      <div @click="gotoList(item)" class="flex">
-        <sub-menu-item :item="item" :icon-size="iconSize" />
-      </div>
-    </template>
-    <el-menu-item
-      @click="gotoList(submenu)"
-      :index="submenu.title"
-      v-for="submenu in item.children"
+  <div v-if="item?.status">
+    <el-sub-menu
+      :index="item.title"
+      v-if="item?.children && item.children.length > 0"
     >
-      <sub-menu-item :item="submenu" :icon-size="iconSize" />
+      <template #title>
+        <div @click="gotoList(item)" class="flex">
+          <sub-menu-item :item="item" :icon-size="iconSize" />
+        </div>
+      </template>
+      <el-menu-item
+        @click="gotoList(submenu)"
+        :index="submenu.title"
+        v-for="submenu in item.children"
+      >
+        <sub-menu-item :item="submenu" :icon-size="iconSize" />
+      </el-menu-item>
+    </el-sub-menu>
+    <el-menu-item @click="gotoList(item)" v-else :index="item?.title">
+      <sub-menu-item :item="item" :icon-size="iconSize" />
     </el-menu-item>
-  </el-sub-menu>
-  <el-menu-item @click="gotoList(item)" v-else :index="item?.title">
-    <sub-menu-item :item="item" :icon-size="iconSize" />
-  </el-menu-item>
+  </div>
 </template>
 
 <style scoped></style>
